+++ CHANGELOG.md
@@ -1,7 +1,10 @@
 # Unreleased
+- [...]
+
+# v0.3.0 (08/06/2018)
 Breaking changes:
 - [Autocomplete] Render empty search with an autocomplete list
-- [Autocomplete] Replace the props `loadingItemIndex` and `valid` by a new selectedItemStatus
+- [Autocomplete] Replace the props `loadingItemIndex` and `valid` by a new selectedItemStatus props
 
 # v0.2.2 (07/06/2018)
 - [Autocomplete] Pass down the `bodyClassName` to the AutoCompleteList